kokoro/requirements.txt
# FastAPI 及 Web 框架
fastapi
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
email-validator

# 数据库及缓存
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.3

# 核心依赖
numpy>=2.0.1,<3.0.0
torch  # macOS ARM64 推荐对应 wheel

# AI 相关
transformers
diffusers
accelerate
peft==0.7.1
sentence-transformers==2.2.2
huggingface-hub
datasets
bittensor==9.12.2

# 文本处理
ftfy==6.1.1
regex==2023.10.3

# 图像处理 / 科学计算
pillow==10.1.0
scipy
scikit-learn

# 其他工具
python-dotenv==1.0.0
gitpython==3.1.40
pyyaml==6.0.1

# macOS 3.12 兼容性补充
watchdog
pathtools
